+++ main/parse.h
@@ -69,6 +69,7 @@ typedef enum {
 	METHOD_XCMD           = 1 << 2,
 	METHOD_XCMD_AVAILABLE = 1 << 3,
 	METHOD_XPATH          = 1 << 4,
+	METHOD_YAML           = 1 << 5,
 } parsingMethod;
 
 typedef struct {
@@ -136,7 +137,7 @@ struct sParserDefinition {
 	simpleParser parser;           /* simple parser (common case) */
 	rescanParser parser2;          /* rescanning parser (unusual case) */
 	selectLanguage* selectLanguage; /* may be used to resolve conflicts */
-	unsigned int method;           /* See PARSE__... definitions above */
+	unsigned int method;           /* See METHOD_ definitions above */
 	bool useCork;
 	bool allowNullTag;
 	bool requestAutomaticFQTag;
@@ -196,6 +197,10 @@ extern parserDefinitionFunc PARSER_LIST;
 #ifdef HAVE_LIBXML
 extern parserDefinitionFunc XML_PARSER_LIST;
 #endif
+#ifdef HAVE_LIBYAML
+extern parserDefinitionFunc YAML_PARSER_LIST;
+#endif
+
 
 /* Language processing and parsing */
 extern int makeSimpleTag (const vString* const name, kindOption* const kinds, const int kind);